bouncer/chainflip-io/packages/swap/python-client/requirements.txt
aiohttp==3.8.4
python-socketio==5.8.0
cryptography==40.0.2
